+++ doc/whats-new.rst
@@ -23,6 +23,8 @@ v2024.03.0 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
+- Allow control over padding in rolling. (:issue:`2007`, :pr:`5603`).
+  By `Kevin Squire <https://github.com/kmsquire>`_.
 - Do not broadcast in arithmetic operations when global option ``arithmetic_broadcast=False``
   (:issue:`6806`, :pull:`8784`).
   By `Etienne Schalk <https://github.com/etienneschalk>`_ and `Deepak Cherian <https://github.com/dcherian>`_.